refactor: [WIP] place tag functions into hook
